# api/swagger.json
{
  "info": {
    "title": "test api",
    "version": "1.0.0",
    "description": "API for test task"
  },
  "schemes": [
    "http"
  ],
  "host": "127.0.0.1:3000",
  "basePath": "/",
  "swagger": "2.0",
  "paths": {
...
    "/provider/{id}": {
      "delete": {
        "tags": [
          "provider"
        ],
        "description": "Deletes provider by id",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "description": "ID of the provider to delete"
          }
        ],
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "Successeful delete"
          },
          "400": {
            "description": "Invalid ID"
          }
        }
      },
      "get": {
        "tags": [
          "provider"
        ],
        "description": "Get provider by id",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "description": "ID of the provider to get"
          }
        ],
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "Provider object",
            "schema": {
              "type": "object",
              "properties": {
                "provider": {
                  "$ref": "#/definitions/Provider"
                }
              }
            }
          },
          "400": {
            "description": "Invalid ID"
          }
        }
      },
      "patch": {
        "tags": [
          "provider"
        ],
        "description": "Edit user by id",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "description": "ID of the client to get"
          },
          {
            "in": "body",
            "name": "body",
            "description": "Provider object with edited fields",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Provider"
            }
          }
        ],
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "Edited provider",
            "schema": {
              "type": "object",
              "properties": {
                "provider": {
                  "$ref": "#/definitions/Provider"
                }
              }
            }
          },
          "400": {
            "description": "Invalid ID"
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "Provider": {
      "type": "object",
      "properties": {
        "id": {
          "type": "number"
        },
        "name": {
          "type": "string"
        }
      }
    }
  },
...
  "securityDefinitions": {},
...
}